fix: fixed up the bugs with button actions

+++ src/main/java/com/seailz/discordjv/utils/discordapi/RateLimit.java
@@ -1,8 +1,8 @@
 package com.seailz.discordjv.utils.discordapi;
 
 public record RateLimit(
-        int limit,
-        int remaining,
-        int resetAfter
+        double limit,
+        double remaining,
+        double resetAfter
 ) {
 }
